requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
selenium==4.26.1
webdriver-manager==4.0.2
cerebras_cloud_sdk==1.11.0
textblob
matplotlib
wordcloud